Define __all__ for all modules
Defining `__all__` in each module makes it explicit what names are
intended to be used by client modules. An "import *" from a module
will only import names that included in `__all__`, if it's defined.

IDEs can warn you if you try to explicitly import something not
included in a module's `__all__`.