Replace "import *" with explicit imports
Using explicit, individual imports instead of "from X import *" makes it
clear to contributors (and others reading the code) which module
each name is being imported from, without relying on an IDE to tell you.

None of the imports in __init__.py are now being relied on, in the
source. The unit tests still rely on them. Ideally they wouldn't,
as this would make the unit tests similarly clearer, but I don't want
to touch the unit tests in this commit, or really in this PR at all.