Auto-generate FFI bindings with `ctypes`
This commit migrates this binding of Wasmtime to auto-generate `ctypes`
definitions for all exported functions and types from wasmtime. The goal
of this is to automate binding generation and avoid having to hand-write
all imported types/functions. Currently all our functions stylistically
specify their return types (some for correctness), and this makes it so
they're procedurally generated and additionally specify argument types.

This should enable us to get better type checking with `ctypes` itself
as well as enable type annotations in this library since all imported
functions are procedurally generated and typed accordingly.