Update to Wasmtime's new C API
This commit updates wasmtime-py to account for [RFC 11] and the changes
in Wasmtime's own C API. Note that this is a breaking change for this
binding, notably requiring that `Store` is a uniquely-owned value now
which must be passed in as context to most methods that operate with a
store. More details are in the commit itself and updated tests and such.

[RFC 11]: https://github.com/bytecodealliance/rfcs/pull/11