Add basic scaffolding for bindgen tests
For now settle on the scheme of a `tests/codegen/test_*` suite of files
which, when imported, will generated bindings on-the-fly which means
that mypy type-checking, testing, etc, can all get folded into a single
`pytest` with good error reporting. All that's left now is to fill out
all the tests for all the various features.