Enable explicit destruction of all resources
This commit adds a suite of methods to most of the types in this package
to enable explicit destruction of resources as necessary. Specifically
most classes which own a pointer internally now sport some new methods:

* `__enter__` and `__exit__` enable using the objects in `with` blocks.
* `close` enables explicitly deallocating the object.

All objects internally now validate that they have not been closed
before operating, raising a `ValueError` if a closed object is operated
on.