loader: assorted fixes (#222)
* loader: do not prepend custom meta finder.

There's basically no reason to do this as nothing in Python ecosystem
attempts to load `.wasm` or `.wat` files except for wasmtime, and it is
[recommended] in the docs to use `.append()` for adding a finder.

[recommended]: https://docs.python.org/3/library/importlib.html#setting-up-an-importer

* loader: simplify. NFC

* _types: fix typo in docstring.

* loader: prettify. NFC

* loader: use slightly nicer partial application.

This approach has the advantage of not having the `func` argument in
the signature of the method, which is visible in e.g. `help()`.

* loader: simplify.

The `spec_from_file_location` helper includes a lot of indirection and
code that we really don't care about; creating the right `ModuleSpec`
directly here makes the semantics obvious.