Initial implemention of bindgen for exported resources (#232)
* Add minimal implementation for WASI stubs

This replaces the `NotImplementedError` with enough of an
implementation to run the bindgen on components with resources.
This gets us further into the bindgen code, which now panics
at the unimplemented parts for resources.

* Initial implemention of bindgen for exported resources

This adds support for generating bindings for resources exported from a
wasm component, so that Python code can interact with these resources.

I looked over jco/componentize-py and followed the same overall
approach.  The internal structuring of `bindgen.rs` here is slightly
different than jco, so I erred on the side of keeping things as they are
as much as possible.

In terms of the generated code itself:

* A resource maps to a Python class.
* Calling the constructor of a resource returns a new instance of a
  Python class.
* Protocols are generated for each resource in the module corresponding
  to the exported interface.
* The protocols follow the existing calling convention of
  `func(caller: Store, ...<modeled args>) for exported functions.
* The instantiation of the resource class is accessible from the instance
  of the class that holds the scope for the exported interface.

One of the issues I had to work out was how to pass in the root object
that contains all the trampolines needed by the resource classes, which
is defined in the root object.  This is an issue because the
`constructor` of a resource is modeled in WIT, and the root object is
typically passed as a param in the `__init__`.  What I ended up doing
was using a closure to capture the root component instance and then
dynamically generating the class when the interface "scope class" is
instantiated.  Would love feedback on this, it's unconventional in
Python, but I suppose it also maps closer to the idea of a `(resource
...)` being a type constructor that creates a fresh type for each
instance, while still being able to use the the module level protocols
for the resource to satisfy type checking of the resources.

Imported resources aren't implemented as part of this change.  Trying to
bindgen imported resources will still result in an `unimplemented!()`
panic.  Wanted to get feedback on the overall approach before going
too far.

* Implement resource.rep canon built-in