Wiring SharedMemory with it's corresponding bindings (#255)
* Wiring SharedMemory with it's corresponding bindings

* Restoring _bindings.py as is being autogenerated out of the C-API
interface in the wasmtime runtime repo

* Removing SharedMemory read/write methods, as they are also not
implemented in the Rust API for wasmtime.
A SharedMemory is not associated to any Store, and the read/write
methods from the linear memory seem  to rely on the Store in order
to properly operate.

* Making SharedMemory inherit from the Manager type

* Adding runtime typechecking to SharedMemory::_from_raw

* After regenerating the bindings from wasmtime c-api (main/dev branch)
and adjusting the shared memory arguments to match the new api, I
found out that I needed to adjust other parts of the bindings that I guess
have changed since the last release.
In this commit, I only change the (shared) memory related ones, but I
think I can adjust the rest of the c-api that has changed as well in
the following commits.

* Update to latest wasmtime dev (#257)

* Retoring bindings from main branch as it was updated to latest
changes in the C-APO from Wasmtime runtime repo

* * Fixing Python types as suggested by mypy
* Adding some unit tests for shared memory

* Adding unit tests for shared memory

---------

Co-authored-by: Jesse Rusak <me@jesserusak.com>