Build source tar.gz (sdist) alongside wheels (bdist_wheel)
This is a first pass. Probably enough to build conda packages, but not enough
for pip or uv to build from. I've tried to be minimally invasive

- Keep setuptools as the packaging backend
- Keep `setup.py` as the source of truth

A more complete overhaul might consider moving to a more recent/focused build
library (e.g. PyO3, Maturin); and/or a more recent (e.g. `setup.cfg`,
`pyproject.toml`).

A diff of file listings before and after show the wheel contents are unchanged,
except for version specific paths

```console
$ diff -u \
  <(zipinfo -1 dist-before/wasmtime-33.0.0.dev362-py3-none-macosx_11_0_arm64.whl|sort) \
  <(zipinfo -1 dist.after/wasmtime-33.0.0.dev363-py3-none-macosx_11_0_arm64.whl|sort)
--- /dev/fd/11  2025-06-10 09:53:26
+++ /dev/fd/12  2025-06-10 09:53:26
@@ -1,8 +1,8 @@
-wasmtime-33.0.0.dev362.dist-info/LICENSE
-wasmtime-33.0.0.dev362.dist-info/METADATA
-wasmtime-33.0.0.dev362.dist-info/RECORD
-wasmtime-33.0.0.dev362.dist-info/top_level.txt
-wasmtime-33.0.0.dev362.dist-info/WHEEL
+wasmtime-33.0.0.dev363.dist-info/LICENSE
+wasmtime-33.0.0.dev363.dist-info/METADATA
+wasmtime-33.0.0.dev363.dist-info/RECORD
+wasmtime-33.0.0.dev363.dist-info/top_level.txt
+wasmtime-33.0.0.dev363.dist-info/WHEEL
 wasmtime/__init__.py
 wasmtime/_bindings.py
 wasmtime/_config.py
```

refs #231