Build source distributions (sdist) alongside wheels (bdist_wheel)
I have integrated running of the ci/download-wasmtime.py and ci/build-rsyt.py
into the PEP-517 `build_wheel()` step. This allows a wheel to be built from
just a source distribution (as opposed to a git checkout), so e.g.

```
pip install --no-binary=:any: wasmtime-py
```

has a chance of working.

The integration mechanism is ci/_custom_build/backend.py, an in-tree PEP-517
build backend that extends the build backend provided by setuptools. This file
is excluded from pytest discovery/scanning for expediency and to keep the
scope of this change smaller. A future PR could remove this restriction to
improve test and style guide coverage.

I elected to keep most project metadata in setup.py for now, again to minimize
the size of the PR. I recommend a future PR to move the metadata to
pyproject.toml.

This change incorporates the changes begun in PR #285, and supercedes it. As
with that PR wheels builtin in wasmtime-pr should be unchanged - so in the
common case end users do not require a rust compiler, or to build from source.