Use Python "build" package as build frontend, don't directly run setup.py (#287)
* Use Python "build" package as build frontend, don't directly run setup.py

Running setup.py directly is deprecated, from 2025-10-31 it won't be
supported. The Python ecosystem is moving toward seperation of build frontend
and backend. This commit switches to the "build" frontend, and keeps
setuptools as the backend.

Two `setup()` keywords have been altered

`packages` was incomplete, but happened to work. Setuptools now warns about
such a condition, which added noise in the build output.

`test_suite` was unnnecessary because pytest is used for testing. It has been
deprecated since by setuptools 41.5.0 (Oct 2019), and its presence emitted a
warning that added noise to build output.

This change should
- allow wasmtime-py wheels to be built, published, and installed.

It should not
- alter versioning behaviour (specifically .devNNN tagging)
- alter wheel contents

It does not
- allow building of an sdist (source tar.gz distribution)

Risks
- `-C=--build-option=--plat-name -C=--build-option=...` is ugly and
  possibly fragile. A future release of build or setuptools may break it.
- No version constraints are placed on setuptools in pyproject.toml.

I think this is the minimum viable change that fixes #286. Other improvements
can be made seperately (e.g. migrating more metadata from setup.py to
pyproject.toml).

See
- https://build.pypa.io/en/stable/
- https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
- https://setuptools.pypa.io/en/latest/references/keywords.html#keyword-test-suite
- https://github.com/pypa/pip/issues/11859
- https://github.com/pypa/setuptools/issues/2491
- https://github.com/pypa/setuptools/discussions/3882

* Revert change to packages keyword in setup.py

CI was failing because the "Setup wasmtime-py CI" action in .github/actions/
setup.yml installs `.[testing]` before ci/build-rust.py has been run. Thus
wasmtime/bindgen/generated and wasmtime/bindgen/generated/imports didn't exist
yet.

See
- https://github.com/bytecodealliance/wasmtime-py/actions/runs/15640817846/job/44067456534

* Build source distributions (sdist) alongside wheels (bdist_wheel)

I have integrated running of the ci/download-wasmtime.py and ci/build-rsyt.py
into the PEP-517 `build_wheel()` step. This allows a wheel to be built from
just a source distribution (as opposed to a git checkout), so e.g.

```
pip install --no-binary=:any: wasmtime-py
```

has a chance of working.

The integration mechanism is ci/_custom_build/backend.py, an in-tree PEP-517
build backend that extends the build backend provided by setuptools. This file
is excluded from pytest discovery/scanning for expediency and to keep the
scope of this change smaller. A future PR could remove this restriction to
improve test and style guide coverage.

I elected to keep most project metadata in setup.py for now, again to minimize
the size of the PR. I recommend a future PR to move the metadata to
pyproject.toml.

This change incorporates the changes begun in PR #285, and supercedes it. As
with that PR wheels builtin in wasmtime-pr should be unchanged - so in the
common case end users do not require a rust compiler, or to build from source.