Use setuptools-git-versioning for dist versions (e.g. 1.2.3.dev12)
This change restores a `.devN` suffix to the version of distributions built
from an untagged commit. The suffix indicates they're a pre-release and makes
the version unique - so the dists can be automatically published to
test.py.pi.org.

Tagged commits will produce a version without any suffix, then be published to
both test.pypi.org and pypi.org.

The change replaces a previous mechanism that was disabled in #287, while
migrating CI to a PEP 517 build frontend/backend.

Differences
- Target version bumped from 33.0.0 -> 33.0.1
- Target version now stored in the file VERSION
- In ".devN" N now counts commits since the file VERSION was modified,
  previously N counted all commits back to the initial commit
- Distributions built from a dirty working tree will generate a version such
  as "1.2.3.post0+git.45fd99a2.dirty" using `dirty_template` config
- The `upload_dists` CI job no longer does a checkout or run of the project
  setup action. They should be unnecessary, as `setup.py` no longer runs `git`
- setuptools-git-versioning is a new build time dependency, it will be
  installed by any PEP 517 build frontend (e.g. python -m build, pip, uv)

Fixes #288

See
- https://setuptools-git-versioning.readthedocs.io
- https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
- https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
- https://peps.python.org/pep-0517/