Remove importlib_resources dependency
This replaces a single use of importlib_resources.files() with
importlib.resources.files() which was introduced in Python 3.9. Now that
wasmtime-py requires Python >= 3.9 the compatibility shim should no longer be
necessary.

importlib_resources (with an underscore) is a Python package that backports
newer parts of importlib.resources (with a .) to older Python releases.