Fix ResourceReader ImportError exception on Python 3.14 (#299)
importlib.abc.ResourceReader moved to importlib.resources.abc.ResourceReader
in Python 3.11. The old location was deprecated in 3.11 and removed in 3.14.

TraversableResources is now the preferred API, but AIUI ResourceReader has no
formal deprecation at the new location. I've chosen an if/else block because
it is interpretted by MyPy, as opposed to try/except ImportError which MyPy
cannot type check on Python <= 3.10.